Fix farm crash when account was not connected. Add address fallback for unconnected accounts in hook. Add rendering portfolio component when account is connected
